+++ packages/tests/lib.esm/test-providers.js
@@ -405,6 +405,7 @@ function testProvider(providerName, networkName) {
     // Delay (ms) after each test case to prevent the backends from throttling
     const delay = 1000;
     describe(("Read-Only " + providerName + " (" + networkName + ")"), function () {
+        this.retries(3);
         // Get the Provider based on the name of the provider we are testing and the network
         let provider = null;
         if (networkName === "default") {
@@ -585,7 +586,7 @@ describe("Test extra Etherscan operations", function() {
 });
 */
 describe("Test Basic Authentication", function () {
-    // https://stackoverflow.com/questions/6509278/authentication-test-servers#16756383
+    this.retries(3);
     function test(name, url) {
         it("tests " + name, function () {
             this.timeout(60000);
@@ -685,6 +686,7 @@ describe("Test API Key Formatting", function () {
     });
 });
 describe("Test WebSocketProvider", function () {
+    this.retries(3);
     function testWebSocketProvider(provider) {
         return __awaiter(this, void 0, void 0, function* () {
             yield provider.destroy();
@@ -698,6 +700,7 @@ describe("Test WebSocketProvider", function () {
     });
 });
 describe("Test Events", function () {
+    this.retries(3);
     function testBlockEvent(provider) {
         return __awaiter(this, void 0, void 0, function* () {
             return new Promise((resolve, reject) => {